# Allow arbitrary integers in take, drop, repeated, and partition