Fixes #19214

Previously `take`, `drop`, `repeated`, and `partition` only allowed `Int` arguments. This widens the signatures to permit any `Integer` type.