@ararslan 2016-11-10
The Linux failures appear to be unrelated; something about spawn. Doesn't happen on Windows or OS X.