@martinholters 2016-11-11
Have you checked performance? The `n::Integer` fields seem dangerous in this respect. Maybe better have `n::Tn` with `Tn<:Integer` being a parameter of the enclosing type?