@ararslan 2016-11-11
@martinholters That's a good point, thanks. I've run some benchmarks using BenchmarkTools and the change appears not to affect performance; the results are almost identical. I can still add a parameter to the type if you think that's better though.