@ararslan 2016-11-11
@yuyichao For example,
```julia
julia> @benchmark Iterators.take(rand(10_000_000), 5) # Built from master
BenchmarkTools.Trial:
  memory estimate:  76.29 mb
  allocs estimate:  3
  --------------
  minimum time:     25.808 ms (24.35% GC)
  median time:      26.486 ms (24.11% GC)
  mean time:        29.127 ms (24.83% GC)
  maximum time:     98.446 ms (79.61% GC)
  --------------
  samples:          172
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Iterators.take(rand(10_000_000), 5) # Built on this branch
BenchmarkTools.Trial:
  memory estimate:  76.29 mb
  allocs estimate:  3
  --------------
  minimum time:     25.745 ms (24.15% GC)
  median time:      26.123 ms (24.17% GC)
  mean time:        27.223 ms (25.65% GC)
  maximum time:     79.097 ms (74.95% GC)
  --------------
  samples:          184
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```
I ran similar tests for `drop`, `repeated`, and `partition`.