@KristofferC 2016-11-11
Don't you think the `rand(10_000_000)` will dominate the execution time making your benchmark not actually test the performance of `take`?