@ararslan 2016-11-11
@KristofferC That's a very good point, I hadn't considered that. I'll try again using a predefined variable.