@yuyichao 2016-11-11
Also the creation of the iteration is not he problem at all, the use of it is.