@ararslan 2016-11-11
@yuyichao What would you suggest as a more appropriate benchmark?