@yuyichao 2016-11-11
```
julia> using BenchmarkTools
INFO: Recompiling stale cache file /home/yuyichao/.julia/lib/v0.6/JLD.ji for module JLD.

julia> v = Vector{Int}(1000_000);

julia> it = Iterators.take(Iterators.repeated(1000_000), 1000_000);

julia> @benchmark copy!($v, $it)
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     4.165 ms (0.00% GC)
  median time:      4.608 ms (0.00% GC)
  mean time:        4.557 ms (0.00% GC)
  maximum time:     7.104 ms (0.00% GC)
  --------------
  samples:          1097
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Iterators.take(Iterators.repeated(1000_000), 1000_000)
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.629 ns (0.00% GC)
  median time:      2.088 ns (0.00% GC)
  mean time:        2.025 ns (0.00% GC)
  maximum time:     13.727 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```