@yuyichao 2016-11-11
And on this branch

```jl
julia> v = Vector{Int}(1000_000);

julia> it = Iterators.take(Iterators.repeated(1000_000), 1000_000);

julia> @benchmark copy!($v, $it)
BenchmarkTools.Trial: 
  memory estimate:  61.04 mb                                                      
  allocs estimate:  3000001                                                       
  --------------                                                                  
  minimum time:     84.246 ms (5.11% GC)                                          
  median time:      92.733 ms (7.26% GC)                                          
  mean time:        96.707 ms (6.91% GC)                                          
  maximum time:     156.786 ms (5.34% GC)                                         
  --------------                                                                  
  samples:          52                                                            
  evals/sample:     1                                                             
  time tolerance:   5.00%                                                         
  memory tolerance: 1.00%                                                         
                                                                                  
julia> @benchmark Iterators.take(Iterators.repeated(1000_000), 1000_000)
BenchmarkTools.Trial:                                                             
  memory estimate:  32.00 bytes                                                   
  allocs estimate:  1                                                             
  --------------                                                                  
  minimum time:     6.728 ns (0.00% GC)                                           
  median time:      9.169 ns (0.00% GC)                                           
  mean time:        11.043 ns (15.39% GC)                                         
  maximum time:     1.307 μs (98.20% GC)                                          
  --------------                                                                  
  samples:          10000                                                         
  evals/sample:     1000                                                          
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

So unsurprisingly ~20x and ~4.5x slower on average.