@ararslan 2016-11-11
Ah dang, that's quite unfortunate. Thanks so much for running those benchmarks! You beat me to it. I'll update the branch as per @martinholters' suggestion and see how that affects things.