@yuyichao 2016-11-12
Also note that  #19214 does not require changing the field type.