@ararslan 2016-11-12
On this branch, updated to use static parameters:

```julia
julia> @benchmark copy!($v, $it)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     4.950 ms (0.00% GC)
  median time:      5.142 ms (0.00% GC)
  mean time:        5.430 ms (0.00% GC)
  maximum time:     11.570 ms (0.00% GC)
  --------------
  samples:          920
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Iterators.take(Iterators.repeated(1000_000), 1000_000)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.086 ns (0.00% GC)
  median time:      2.094 ns (0.00% GC)
  mean time:        2.161 ns (0.00% GC)
  maximum time:     10.094 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

On master:

```julia
julia> @benchmark copy!($v, $it)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     4.952 ms (0.00% GC)
  median time:      5.159 ms (0.00% GC)
  mean time:        5.420 ms (0.00% GC)
  maximum time:     11.152 ms (0.00% GC)
  --------------
  samples:          922
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Iterators.take(Iterators.repeated(1000_000), 1000_000)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.087 ns (0.00% GC)
  median time:      2.096 ns (0.00% GC)
  mean time:        2.375 ns (0.00% GC)
  maximum time:     27.966 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```