@ararslan 2016-11-12
@yuyichao Yeah, I suppose another option would be to use `Integer` in the method signature and just convert the argument to `Int` when constructing the actual `Take`/`Drop`/etc. object. I assume that's what you meant? That would probably be the least disruptive change.